+++ HPSpellBook/package.json
@@ -17,7 +17,8 @@
     "react": "^18.0.0",
     "react-native": "0.69.4",
     "react-native-safe-area-context": "4.3.1",
-    "react-native-screens": "~3.15.0"
+    "react-native-screens": "~3.15.0",
+    "react-native-uuid": "^2.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
